chore: clean up (3)
