+++ apps/www/package.json
@@ -64,9 +64,7 @@
     "tailwind-variants": "0.2.1",
     "ts-pattern": "5.4.0",
     "zod": "3.23.8",
-    "zod-formik-adapter": "1.3.0",
-    "zsa": "0.6.0",
-    "zsa-react": "0.2.3"
+    "zod-formik-adapter": "1.3.0"
   },
   "devDependencies": {
     "@1.config/tailwindcss": "workspace:*",
