+++ packages/@1/modules/auth/next/src/actions/signup.ts
@@ -8,9 +8,9 @@ import {
   NewStudent_Schema,
 } from "@1.modules/profile.domain";
 import { gravatarUrlFor } from "@1.modules/profile.domain/gravatarUrlFor";
+import { createServerActionProcedure } from "@1.ui/react/zsa";
 import { match } from "ts-pattern";
 import z from "zod";
-import { createServerActionProcedure } from "zsa";
 import { auth, signIn } from "../auth.config";
 
 const is_auth_user_procedure = createServerActionProcedure().handler(
