--- packages/@1/modules/auth/next/src/csrf_token.ts
@@ -1,3 +1,0 @@
-//
-
-export { get_csrf_token } from "@douglasduteil/nextauth...trpc.prisma/next/csrf_token";
