+++ packages/@1/modules/core/src/errors.ts
@@ -4,7 +4,7 @@ import errorCustomClass from "error-custom-class";
 
 //
 
-export const HTTPError = errorCustomClass("HTTPError");
 export const AuthError = errorCustomClass("AuthError");
+export const HTTPError = errorCustomClass("HTTPError");
 export const ProfileNotFoundError = errorCustomClass("ProfileNotFoundError");
 export const StateError = errorCustomClass("StateError");
