+++ packages/@1/modules/trpc/src/index.ts
@@ -19,11 +19,7 @@ export {
   router,
 } from "./trpc";
 
-export {
-  maybe_next_auth_procedure,
-  next_auth_input_token,
-  next_auth_procedure,
-} from "./guards";
+export { maybe_next_auth_procedure, next_auth_procedure } from "./guards";
 
 export type inferProcedureInput<TQuery extends ProcedureBuilder<any>> =
   TQuery extends ProcedureBuilder<infer TParams>
