+++ packages/@1/ui/react/package.json
@@ -53,6 +53,10 @@
       "types": "./src/useragent/index.ts",
       "default": "./src/useragent/index.ts"
     },
+    "./zsa": {
+      "types": "./src/zsa/index.ts",
+      "default": "./src/zsa/index.ts"
+    },
     "./zsa/react": {
       "types": "./src/zsa/react.ts",
       "default": "./src/zsa/react.ts"
