+++ packages/@1/ui/react/src/zsa/react-query.ts
@@ -5,17 +5,6 @@ import { setupServerActionHooks } from "zsa-react-query";
 
 //
 
-// export const QueryKeyFactory = createServerActionsKeyFactory({
-//   // [!code highlight]
-//   getPosts: () => ["getPosts"], // [!code highlight]
-//   getFriends: () => ["getFriends"], // [!code highlight]
-//   getPostsAndFriends: () => ["getPosts", "getFriends"], // [!code highlight]
-//   somethingElse: (id: string) => ["somethingElse", id], // [!code highlight]
-//   getRandomNumber: () => ["getRandomNumber"], // [!code highlight]
-// }); // [!code highlight]
-
-//
-
 export const {
   useServerActionQuery,
   useServerActionMutation,
@@ -26,5 +15,4 @@ export const {
     useMutation: useMutation,
     useInfiniteQuery: useInfiniteQuery,
   },
-  // queryKeyFactory: QueryKeyFactory,
 });
