+++ packages/@douglasduteil/nextauth...trpc.prisma/package.json
@@ -22,22 +22,11 @@
       "default": "./src/*.ts"
     }
   },
-  "main": "./src/index.ts",
-  "types": "./src/index.ts",
   "scripts": {
-    "_generate": "prisma generate",
-    "dev:studio": "prisma studio",
-    "format": "prisma format",
-    "migrate:deploy": "prisma migrate deploy",
-    "migrate:dev": "prisma migrate dev",
-    "push": "prisma db push --skip-generate",
-    "seed": "prisma db seed",
-    "studio": "prisma studio",
     "type-check": "tsc --noEmit"
   },
   "prettier": "@1.config/prettier",
   "dependencies": {
-    "@prisma/client": "5.20.0",
     "@trpc/client": "10.45.2",
     "@trpc/server": "10.45.2",
     "next": "14.2.13",
