+++ packages/@douglasduteil/nextauth...trpc.prisma/src/jwt.ts
@@ -7,7 +7,6 @@ import {
   type JWT as NextAuth_JWT,
   type JWTEncodeParams as NextAuth_JWTEncodeParams,
 } from "next-auth/jwt";
-import { createHash } from "node:crypto";
 
 export interface JWT extends NextAuth_JWT {
   from?: string;
@@ -25,20 +24,6 @@ export async function create_nextauth_header<TToken extends object>(
   return { NEXTAUTH_TOKEN: await jwt_encode(params) };
 }
 
-/** Decodes a NextAuth.js issued JWT. */
 export async function decode(params: JWTDecodeParams): Promise<JWT | null> {
   return (await jwt_decode(params)) as JWT;
-}
-
-export function hashToken(
-  token: string,
-  options: { provider: { secret?: string }; secret?: string },
-) {
-  const { provider, secret } = options;
-  return (
-    createHash("sha256")
-      // Prefer provider specific secret, but use default secret if none specified
-      .update(`${token}${provider.secret ?? secret}`)
-      .digest("hex")
-  );
 }
