--- packages/@douglasduteil/nextauth...trpc.prisma/src/next/index.ts
@@ -1,4 +1,0 @@
-//
-
-export { PrismaTRPCAdapter } from "./adapter";
-export { get_csrf_token } from "./csrf_token";
