--- packages/@douglasduteil/nextauth...trpc.prisma/src/trpc/index.ts
@@ -1,3 +1,0 @@
-//
-
-export * from "./guards";
